+++ nodalia-cards-2.1.3-alpha.2.js
@@ -1,2 +1,2 @@
-import "./nodalia-cards-2.1.3-alpha.7.js";
-if(typeof window!=="undefined"){window.__NODALIA_LOADER__={"mode":"compat","pkgVersion":"2.1.3-alpha.7","contentSha256_12":"73be3c9b6b07","file":"nodalia-cards-2.1.3-alpha.2.js","targetFile":"nodalia-cards-2.1.3-alpha.7.js","fallbackFile":"nodalia-cards.js"};}
+import "./nodalia-cards-2.1.3-alpha.8.js";
+if(typeof window!=="undefined"){window.__NODALIA_LOADER__={"mode":"compat","pkgVersion":"2.1.3-alpha.8","contentSha256_12":"b3ab73ba694f","file":"nodalia-cards-2.1.3-alpha.2.js","targetFile":"nodalia-cards-2.1.3-alpha.8.js","fallbackFile":"nodalia-cards.js"};}