+++ nodalia-cards.manifest.js
@@ -1,6 +1,6 @@
 export default {
-  "pkgVersion": "2.1.3-alpha.7",
-  "contentSha256_12": "73be3c9b6b07",
+  "pkgVersion": "2.1.3-alpha.8",
+  "contentSha256_12": "b3ab73ba694f",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
   "hacsFile": "nodalia-cards.js",
@@ -8,16 +8,16 @@ export default {
     "nodalia-cards-2.1.3-alpha.2.js",
     "nodalia-cards-2.1.3-alpha.3.js"
   ],
-  "splitCoreFile": "nodalia-cards-core-2.1.3-alpha.7.js",
+  "splitCoreFile": "nodalia-cards-core-2.1.3-alpha.8.js",
   "splitCoreSha256_12": "86f0981ace58",
-  "splitSuiteFile": "nodalia-cards-suite-2.1.3-alpha.7.js",
-  "splitSuiteSha256_12": "06bb670b6fc4",
-  "editorFile": "nodalia-cards-editor-2.1.3-alpha.7.js",
+  "splitSuiteFile": "nodalia-cards-suite-2.1.3-alpha.8.js",
+  "splitSuiteSha256_12": "eea8f5facfae",
+  "editorFile": "nodalia-cards-editor-2.1.3-alpha.8.js",
   "editorSha256_12": "910d4e2d8ad0"
 };
-export const pkgVersion = "2.1.3-alpha.7";
-export const contentSha256_12 = "73be3c9b6b07";
+export const pkgVersion = "2.1.3-alpha.8";
+export const contentSha256_12 = "b3ab73ba694f";
 export const file = "nodalia-cards.bundle.js";
-export const splitCoreFile = "nodalia-cards-core-2.1.3-alpha.7.js";
-export const splitSuiteFile = "nodalia-cards-suite-2.1.3-alpha.7.js";
-export const editorFile = "nodalia-cards-editor-2.1.3-alpha.7.js";
+export const splitCoreFile = "nodalia-cards-core-2.1.3-alpha.8.js";
+export const splitSuiteFile = "nodalia-cards-suite-2.1.3-alpha.8.js";
+export const editorFile = "nodalia-cards-editor-2.1.3-alpha.8.js";